Mavenize and support material IDs in builders.
Rearranged project and added Maven POM file so it can be built with Maven.
Added support for specifying material data IDs in builders, e.g. 95:15 for "black stained glass" (but not for "replace" material yet).
