+++ src/main/java/me/simplex/buildr/util/Buildr_Type_Wall.java
