+++ src/main/resources/plugin.yml
